refactor(ui/input): use inputBox for wrap component
