+++ packages/varlet-ui/src/input/InputBox.vue
@@ -1,0 +1,190 @@
+<template>
+  <div
+    :class="classes(n(), n(`--${variant}`), [size, n('--small')], n('$--box'), [disabled, n('--disabled')])"
+    @click="handleClick"
+  >
+    <label :class="classes(n('$--relative'), n('$--block'))" :for="id">
+      <div
+        :class="
+          classes(
+            n('controller'),
+            [isFocus, n('--focus')],
+            [errorMessage, n('--error')],
+            [formDisabled || disabled, n('--disabled')]
+          )
+        "
+        :style="{
+          color: !errorMessage ? (isFocus ? focusColor : blurColor) : undefined,
+        }"
+      >
+        <div :class="classes(n('icon'), [!hint, n('--non-hint')])" ref="prependRef">
+          <slot name="prepend-icon" />
+        </div>
+
+        <div :class="classes(n('wrap'), [!hint, n('--wrap-non-hint')])">
+          <slot />
+          <div
+            :class="
+              classes(
+                n('$--ellipsis'),
+                [isFocus, n('--focus')],
+                [formDisabled || disabled, n('--disabled')],
+                [errorMessage, n('--error')],
+                n('placeholder'),
+                computePlaceholderState(),
+                [!hint, n('--placeholder-non-hint')]
+              )
+            "
+            :style="{
+              color: !errorMessage ? (isFocus ? focusColor : blurColor) : undefined,
+            }"
+          >
+            {{ placeholder }}
+          </div>
+        </div>
+
+        <div :class="classes(n('icon'), [!hint, n('--non-hint')])">
+          <slot name="append-icon">
+            <var-icon
+              :class="n('clear-icon')"
+              var-input-box-cover
+              name="close-circle"
+              size="14px"
+              v-if="clearable && !isEmpty(value)"
+              @click="handleClear"
+            />
+          </slot>
+        </div>
+      </div>
+
+      <template v-if="line">
+        <div
+          v-if="variant === 'outlined'"
+          :class="classes(n('line'), [isFocus, n('--line-focus')], [formDisabled || disabled, n('--line-disabled')])"
+        >
+          <template v-if="!(formDisabled || disabled)">
+            <div
+              :class="classes(n('line__start'), [errorMessage, n('--line-error')])"
+              :style="{
+                borderColor: !errorMessage ? (isFocus ? focusColor : blurColor) : undefined,
+              }"
+            />
+            <div
+              :class="
+                classes(
+                  n('line__notch'),
+                  [hint && (!isEmpty(value) || isFocus), n('line__notch--hint')],
+                  [errorMessage, n('--line-error')]
+                )
+              "
+              :style="{
+                borderColor: !errorMessage ? (isFocus ? focusColor : blurColor) : undefined,
+              }"
+            >
+              <div :class="classes(n('line__placeholder'))">{{ placeholder }}</div>
+            </div>
+            <div
+              :class="classes(n('line__end'), [errorMessage, n('--line-error')])"
+              :style="{
+                borderColor: !errorMessage ? (isFocus ? focusColor : blurColor) : undefined,
+              }"
+            />
+          </template>
+        </div>
+
+        <div
+          :class="
+            classes(n('line'), [formDisabled || disabled, n('--line-disabled')], [errorMessage, n('--line-error')])
+          "
+          :style="{ background: !errorMessage ? blurColor : undefined }"
+          v-else
+        >
+          <div
+            :class="
+              classes(
+                n('dot'),
+                [isFocus, n('--line-focus')],
+                [formDisabled || disabled, n('--line-disabled')],
+                [errorMessage, n('--line-error')]
+              )
+            "
+            :style="{ background: !errorMessage ? focusColor : undefined }"
+          />
+        </div>
+      </template>
+    </label>
+
+    <slot name="form-details" />
+  </div>
+</template>
+
+<script lang="ts">
+import VarIcon from '../icon'
+import { defineComponent, ref, onMounted, onUnmounted } from 'vue'
+import { props } from './inputBoxProps'
+import { isEmpty } from '@varlet/shared'
+import { createNamespace, call } from '../utils/components'
+import type { Ref } from 'vue'
+
+const { n, classes } = createNamespace('input-box')
+
+export default defineComponent({
+  name: 'VarInputBox',
+  components: {
+    VarIcon,
+  },
+  props,
+  setup(props) {
+    const prependRef: Ref<HTMLInputElement | null> = ref(null)
+
+    const computePlaceholderState = () => {
+      const { hint, value, isFocus } = props
+      if (!hint && !isEmpty(value)) {
+        return n('--placeholder-hidden')
+      }
+      if (hint && (!isEmpty(value) || isFocus)) {
+        return n('--placeholder-hint')
+      }
+    }
+
+    const handleClear = (e: Event) => {
+      call(props.onClear, e)
+    }
+
+    const handleClick = (e: Event) => {
+      call(props.onClick, e)
+    }
+
+    let observer: MutationObserver | null = null
+    onMounted(() => {
+      const config = { attributes: true, subtree: true }
+      observer = new MutationObserver(() => {
+        console.log(1)
+        observer!.observe(prependRef.value!, config)
+      })
+    })
+
+    onUnmounted(() => {
+      observer!.disconnect()
+    })
+
+    return {
+      prependRef,
+      computePlaceholderState,
+      n,
+      classes,
+      isEmpty,
+      handleClear,
+      handleClick,
+      focus,
+      blur,
+    }
+  },
+})
+</script>
+
+<style lang="less">
+@import '../styles/common';
+@import '../icon/icon';
+@import './inputBox';
+</style>
