+++ packages/varlet-ui/src/select/select.less
@@ -1,21 +1,10 @@
 :root {
-  --select-select-text-color: #555;
-  --select-select-padding: 5px 0;
-  --select-error-color: var(--color-danger);
-  --select-blur-color: #888;
-  --select-focus-color: var(--color-primary);
   --select-scroller-background: #fff;
   --select-scroller-padding: 6px 0;
   --select-scroller-margin: 8px 0 0 0;
   --select-scroller-max-height: 278px;
-  --select-placeholder-size: 16px;
-  --select-icon-padding: 16px 0 0;
-  --select-icon-size: 20px;
   --select-chip-margin: 5px 5px 0;
-  --select-line-size: 1px;
-  --select-line-spread-size: 2px;
   --select-arrow-size: 20px;
-  --select-disabled-color: var(--color-text-disabled);
 }
 
 .var {
@@ -32,15 +21,6 @@
 }
 
 .var-select {
-  width: 100%;
-  color: var(--select-select-text-color);
-
-  &__controller {
-    width: 100%;
-    display: flex;
-    position: relative;
-  }
-
   &__menu[var-select-cover] {
     flex-grow: 1;
     background: transparent;
@@ -64,26 +44,6 @@
     display: flex;
     flex-direction: column;
     width: 100%;
-    padding-top: var(--select-placeholder-size);
-  }
-
-  &__icon {
-    display: flex;
-    align-items: center;
-    padding: var(--select-icon-padding);
-    font-size: 20px;
-  }
-
-  &__placeholder {
-    position: absolute;
-    top: 50%;
-    left: 0;
-    width: 100%;
-    transform-origin: left;
-    transition: top 0.3s, transform 0.3s, width 0.3s, color 0.25s;
-    transform: translate(0, calc(-50% + var(--select-placeholder-size) / 2)) scale(1);
-    font-size: var(--select-placeholder-size);
-    color: var(--select-blur-color);
   }
 
   &__select {
@@ -91,20 +51,14 @@
     justify-content: space-between;
     align-items: center;
     width: 100%;
-    min-height: 32px;
-    padding: var(--select-select-padding);
+    min-height: 24px;
     outline: none;
     border: none;
     font-size: inherit;
-    color: var(--select-select-text-color);
+    color: var(--input-box-input-text-color);
     word-break: break-all;
     overflow-x: hidden;
     transition: color 0.25s;
-  }
-
-  &__label {
-    display: flex;
-    align-items: center;
   }
 
   &__arrow[var-select-cover] {
@@ -112,64 +66,13 @@
     font-size: var(--select-arrow-size);
   }
 
-  &__line {
-    width: 100%;
-    height: var(--select-line-size);
-    background: var(--select-blur-color);
-    transition: background-color 0.25s;
-  }
-
-  &__dot {
-    width: 100%;
-    height: var(--select-line-spread-size);
-    background: var(--select-focus-color);
-    transform: scaleX(0);
-    transform-origin: center;
-    transition: transform 0.3s var(--cubic-bezier), background-color 0.25s;
-  }
-
-  &--placeholder-hint {
-    top: 0;
-    width: 133.33%;
-    transform: translate(0, 0) scale(0.75);
-  }
-
-  &--placeholder-non-hint {
-    top: 50%;
-    transform: translate(0, -50%) scale(1);
-  }
-
-  &--non-hint {
-    padding-top: 0;
-  }
-
-  &--placeholder-hidden {
-    visibility: hidden;
-  }
-
-  &--focus {
-    color: var(--select-focus-color);
-  }
-
-  &--spread {
-    transform: scaleX(1);
-  }
-
   &--disabled {
-    color: var(--select-disabled-color);
+    color: var(--input-box-disabled-color);
     cursor: not-allowed;
   }
 
   &--error {
-    color: var(--select-error-color);
-  }
-
-  &--line-disabled {
-    background: var(--select-disabled-color);
-  }
-
-  &--line-error {
-    background: var(--select-error-color);
+    color: var(--input-box-error-color);
   }
 
   &--arrow-rotate {
